fix: firebase-messaging-sw 파일 에러 수정
